Performance Tweaks

+++ vhsdecode/utils.py
@@ -9,11 +9,11 @@ from vhsdecode.rust_utils import sosfiltfilt_rust
 SCIPY_1_5_OR_HIGHER = parse(scipy.__version__) >= Version("1.5.0")
 
 
-def gen_wave_at_frequency(frequency, sample_frequency, num_samples, gen_func=np.sin):
+def gen_wave_at_frequency(frequency, sample_frequency, num_samples, gen_func=np.sin, dtype=np.float32):
     """Generate a sine wave with the specified parameters."""
     samples = np.arange(num_samples)
     wave_scale = frequency / sample_frequency
-    return gen_func(2 * np.pi * wave_scale * samples)
+    return gen_func(2 * np.pi * wave_scale * samples, dtype=dtype)
 
 
 def gen_compl_wave_at_frequency(frequency, sample_frequency, num_samples):
